+++ Cpu.h
@@ -11,12 +11,12 @@ class CPU
 {
 private:
   Memory *memory;
-  std::map<std::string, int16_t> registers;
-  int16_t fetch();
-  uint16_t decode(int16_t data) const;
-  uint16_t execute(uint16_t opcode); // will return a binary number representing failure or success
+  std::map<std::string, int8_t> registers;
+  int8_t fetch();
+  uint8_t decode(int8_t data) const;
+  uint8_t execute(uint8_t opcode); // will return a binary number representing failure or success
 
 public:
   CPU(Memory *memory);
-  uint16_t start();
+  uint8_t start();
 };
