+++ Memory.h
@@ -6,6 +6,6 @@ class Memory {
   private:
     uint8_t ram[65536];
   public:
-    int16_t read(uint16_t addr) const;
-    int16_t write(uint16_t addr, int16_t value);
+    int8_t read(uint8_t addr) const;
+    int8_t write(uint8_t addr, int8_t value);
 };
